fix RespInteger to properly read a long and update unit test

+++ src/main/java/org/baylight/redis/protocol/RespInteger.java
@@ -7,7 +7,7 @@ public class RespInteger implements RespValue {
     private final long value;
 
     public RespInteger(BufferedInputLineReader reader) throws NumberFormatException, IOException {
-        this(reader.readInt());
+        this(reader.readLong());
     }
 
     public RespInteger(long value) {
